Reduce overhead to make RPC calls
We can avoid the asyncio.timeout context since we already are
awaiting a future and schedule a timeout on the future that
is being awaited instead. This saves a bit of latency which
helps with being able to connect in time to sleepy
devices for the few seconds they are awake.